pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "org-web"
version = "0.1.0"
description = "FastAPI Web Application for Orgmode editing"
# readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
# authors = [
#     {name = "Your Name", email = "your.email@example.com"},
# ]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "bcrypt",
    "python-multipart",
    "qrcode",
    "isort",
    "pyflakes",
    "orgparse",
    "slixmpp",
    "python-dotenv",
    "python-axolotl-curve25519",
    "doubleratchet",
    "omemo",
]

[project.optional-dependencies]
dev = [
    "black",
    "pyrefly",
]

[project.scripts]
org-web = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "templates/**/*",
    "static/**/*",
]

[tool.hatch.build.targets.sdist]
include = [
    "*.py",
    "templates/**/*",
    "static/**/*",
    # "README.md",
    "pyproject.toml",
]
exclude = [
    "__pycache__",
    "*.pyc",
    ".env",
    ".git",
    ".direnv",
    "subscribers.json",
    "whitelist.json",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
